Actualización de estructura de navegacion
